+++ timeline/tests/test_models.py
@@ -30,7 +30,7 @@ PROJECT_TYPE_PROJECT = SODAR_CONSTANTS['PROJECT_TYPE_PROJECT']
 
 
 # Local constants
-TEST_EXTRA_DATA = {'test_key': 'test_val'}
+EXTRA_DATA = {'test_key': 'test_val'}
 
 
 class TimelineEventMixin:
@@ -131,7 +131,7 @@ class TestTimelineEvent(
             event_name='test_event',
             description='description',
             classified=False,
-            extra_data=TEST_EXTRA_DATA,
+            extra_data=EXTRA_DATA,
         )
 
         self.obj_ref = self.make_object_ref(
@@ -140,7 +140,7 @@ class TestTimelineEvent(
             label='test_label',
             name='test_object_name',
             uuid=self.assignment_owner.sodar_uuid,
-            extra_data=TEST_EXTRA_DATA,
+            extra_data=EXTRA_DATA,
         )
 
     def test_initialization(self):
@@ -154,7 +154,7 @@ class TestTimelineEvent(
             'event_name': 'test_event',
             'description': 'description',
             'classified': False,
-            'extra_data': TEST_EXTRA_DATA,
+            'extra_data': EXTRA_DATA,
             'sodar_uuid': self.event.sodar_uuid,
         }
         self.assertEqual(model_to_dict(self.event), expected)
@@ -168,7 +168,7 @@ class TestTimelineEvent(
             event_name='test_event',
             description='description',
             classified=False,
-            extra_data=TEST_EXTRA_DATA,
+            extra_data=EXTRA_DATA,
         )
         expected = {
             'id': self.event.pk,
@@ -179,7 +179,7 @@ class TestTimelineEvent(
             'event_name': 'test_event',
             'description': 'description',
             'classified': False,
-            'extra_data': TEST_EXTRA_DATA,
+            'extra_data': EXTRA_DATA,
             'sodar_uuid': self.event.sodar_uuid,
         }
         self.assertEqual(model_to_dict(self.event), expected)
@@ -193,7 +193,7 @@ class TestTimelineEvent(
             event_name='test_event',
             description='description',
             classified=False,
-            extra_data=TEST_EXTRA_DATA,
+            extra_data=EXTRA_DATA,
         )
         expected = {
             'id': self.event.pk,
@@ -204,7 +204,7 @@ class TestTimelineEvent(
             'event_name': 'test_event',
             'description': 'description',
             'classified': False,
-            'extra_data': TEST_EXTRA_DATA,
+            'extra_data': EXTRA_DATA,
             'sodar_uuid': self.event.sodar_uuid,
         }
         self.assertEqual(model_to_dict(self.event), expected)
@@ -219,7 +219,7 @@ class TestTimelineEvent(
             event_name='test_event',
             description='description',
             classified=False,
-            extra_data=TEST_EXTRA_DATA,
+            extra_data=EXTRA_DATA,
         )
         expected = {
             'id': self.event.pk,
@@ -230,7 +230,7 @@ class TestTimelineEvent(
             'event_name': 'test_event',
             'description': 'description',
             'classified': False,
-            'extra_data': TEST_EXTRA_DATA,
+            'extra_data': EXTRA_DATA,
             'sodar_uuid': self.event.sodar_uuid,
         }
         self.assertEqual(model_to_dict(self.event), expected)
@@ -356,7 +356,7 @@ class TestTimelineEventObjectRef(
             event_name='test_event',
             description='description',
             classified=False,
-            extra_data=TEST_EXTRA_DATA,
+            extra_data=EXTRA_DATA,
         )
         self.obj_ref = self.make_object_ref(
             event=self.event,
@@ -364,7 +364,7 @@ class TestTimelineEventObjectRef(
             label='test_label',
             name='test_name',
             uuid=self.assignment_owner.sodar_uuid,
-            extra_data=TEST_EXTRA_DATA,
+            extra_data=EXTRA_DATA,
         )
 
     def test_initialization(self):
@@ -376,7 +376,7 @@ class TestTimelineEventObjectRef(
             'name': 'test_name',
             'object_model': 'RoleAssignment',
             'object_uuid': self.assignment_owner.sodar_uuid,
-            'extra_data': TEST_EXTRA_DATA,
+            'extra_data': EXTRA_DATA,
             'sodar_uuid': self.obj_ref.sodar_uuid,
         }
         self.assertEqual(model_to_dict(self.obj_ref), expected)
@@ -417,19 +417,19 @@ class TestTimelineEventStatus(
             event_name='test_event',
             description='description',
             classified=False,
-            extra_data=TEST_EXTRA_DATA,
+            extra_data=EXTRA_DATA,
         )
         self.event_status_submit = self.make_event_status(
             event=self.event,
             status_type='SUBMIT',
             description='SUBMIT',
-            extra_data=TEST_EXTRA_DATA,
+            extra_data=EXTRA_DATA,
         )
         self.event_status_ok = self.make_event_status(
             event=self.event,
             status_type='OK',
             description='OK',
-            extra_data=TEST_EXTRA_DATA,
+            extra_data=EXTRA_DATA,
         )
 
     def test_initialization(self):
@@ -440,7 +440,7 @@ class TestTimelineEventStatus(
             'event': self.event.pk,
             'status_type': 'OK',
             'description': 'OK',
-            'extra_data': TEST_EXTRA_DATA,
+            'extra_data': EXTRA_DATA,
         }
         self.assertEqual(model_to_dict(self.event_status_ok), expected)
 
@@ -452,7 +452,7 @@ class TestTimelineEventStatus(
             'event': self.event.pk,
             'status_type': 'OK',
             'description': 'OK',
-            'extra_data': TEST_EXTRA_DATA,
+            'extra_data': EXTRA_DATA,
         }
         self.event = self.make_event(
             project=self.project,
@@ -461,7 +461,7 @@ class TestTimelineEventStatus(
             event_name='test_event',
             description='description',
             classified=False,
-            extra_data=TEST_EXTRA_DATA,
+            extra_data=EXTRA_DATA,
         )
         self.assertEqual(model_to_dict(self.event_status_ok), expected)
 
@@ -502,7 +502,7 @@ class TestTimelineEventStatus(
             'event': self.event.pk,
             'status_type': 'OK',
             'description': 'OK',
-            'extra_data': TEST_EXTRA_DATA,
+            'extra_data': EXTRA_DATA,
         }
         self.assertEqual(model_to_dict(status), expected)
 
@@ -527,7 +527,7 @@ class TestTimelineEventStatus(
     def test_set_status(self):
         """Test TimelineEventStatus set_status()"""
         new_status = self.event.set_status(
-            'FAILED', status_desc='FAILED', extra_data=TEST_EXTRA_DATA
+            'FAILED', status_desc='FAILED', extra_data=EXTRA_DATA
         )
         expected = {
             'id': new_status.pk,
@@ -535,6 +535,6 @@ class TestTimelineEventStatus(
             'event': self.event.pk,
             'status_type': 'FAILED',
             'description': 'FAILED',
-            'extra_data': TEST_EXTRA_DATA,
+            'extra_data': EXTRA_DATA,
         }
         self.assertEqual(model_to_dict(new_status), expected)